pyproject.toml
[tool.poetry]
name = "napples"
version = "0.1.0"
description = "Backend for Naples Gateway project"
authors = ["Varvara Shcherbyna <varvarashcherbyna7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
alchemical = "^1.0.1"
alembic = "^1.13.1"
python-jose = "^3.3.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.7"
mypy = "^1.9.0"
ipython = "^8.23.0"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
files = ["naples/**/*.py", "tests/**/*.py"]

[tool.ruff]
exclude = [
    "__pypackages__",
    "__pycache__",
    ".eggs",
    ".git",
    ".github",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "migrations",
]
line-length = 120
indent-width = 4
